♻️ : #467 - ProgressBar 컴포넌트 비즈니스 로직 분리
