+++ src/app/(header)/home/_components/hooks/index.ts
@@ -1,1 +1,2 @@
 export { useNewPlan } from './useNewPlan';
+export { default as useProgressBar } from './useProgressBar';
